Change fps and recommended_fps to float

+++ factory-ai-vision/EdgeSolution/modules/WebModule/backend/vision_on_edge/inference_modules/api/serializers.py
@@ -14,7 +14,7 @@ class InferenceModuleSerializer(serializers.ModelSerializer):
     """InferenceModuleSerializer"""
 
     device = serializers.CharField(required=False, read_only=True)
-    recommended_fps = serializers.IntegerField(required=False, read_only=True)
+    recommended_fps = serializers.FloatField(required=False, read_only=True)
 
     class Meta:
         model = InferenceModule
